+++ packages/ui/CHANGELOG.md
@@ -6,7 +6,11 @@ The following is a curated list of changes in the Enact ui module, newest change
 
 ### Added
 
-- `ui/FloatingLayer.FloatingLayerBase` export 
+- `ui/FloatingLayer.FloatingLayerBase` export
+
+### Changed
+
+- `ui/FloatingLayer` to call `onOpen` only after it is rendered
 
 ## [2.0.0-beta.6] - 2018-06-04