+++ packages/ui/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 The following is a curated list of changes in the Enact ui module, newest changes on the top.
 
+## [unreleased]
+
+### Added
+
+- `ui/FloatingLayer.FloatingLayerBase` export 
+
 ## [2.0.0-beta.6] - 2018-06-04
 
 ### Fixed