+++ packages/ui/CHANGELOG.md
@@ -7,6 +7,7 @@ The following is a curated list of changes in the Enact ui module, newest change
 ### Fixed
 
 - `ui/MarqueeDecorator` to stop marqueeing when using hover and pointer hides
+
 ### Added
 
 - `ui/FloatingLayer.FloatingLayerBase` export